src/app/backend/functions/requirements.txt
# Firebase Functions - LiquidChess Discord Bot
# Uses Discord HTTP Interactions for slash commands (serverless-compatible)

firebase-admin>=6.0.0
firebase-functions>=0.1.0
flask>=2.0.0
pynacl>=1.5.0
python-chess>=1.10.0
aiohttp>=3.9.0
requests>=2.31.0
groq>=0.4.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
